fix #451 merge TimedScheduler into Scheduler
This commit merges the functionality of SingleTimedScheduler into its
SingleTimeScheduler sibling, now that both expose an API for delayed
submit.

The 3 task implementations of SingleTimedScheduler and SingleScheduler)
have all been replaced by an extracted ScheduledRunnable.

Time capable schedulers are identifiable through the isTimeCapable
method, while others will immediately return a REJECTED equivalent:
NOT_TIMED.

Time-capable schedulers are:
 - SingleScheduler (now also covering the old SingleTimedScheduler)
 - SingleWorkerScheduler (but not the createWorker() workers)
 - ExecutorServiceScheduler
 - ElasticScheduler
 - ParallelScheduler
 - CachedScheduler (delegates)

isTimeCapable() is used in Flux/Mono APIs that take a Scheduler and
need to schedule in the future, as a mean of failing the operator call
early if the Scheduler provided is not adapted.

Time-scheduling tests have been added to AbstractSchedulerTest and are
only triggered if the scheduler/worker isTimeCapable.

[Behavior Change]
As demonstrated in SchedulersTest, the single scheduler Worker won't
execute an immediate task anymore if it is disposed right after
`submit(task)`.